+++ test/micro/org/openjdk/bench/vm/gc/systemgc/AllLive.java
@@ -20,7 +20,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package org.openjdk.bench.vm.gc.system_gc;
+package org.openjdk.bench.vm.gc.systemgc;
 
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.BenchmarkMode;
@@ -32,35 +32,26 @@ import org.openjdk.jmh.annotations.Scope;
 import org.openjdk.jmh.annotations.Setup;
 import org.openjdk.jmh.annotations.State;
 
-import java.util.HashMap;
+import java.util.ArrayList;
 import java.util.concurrent.TimeUnit;
 
 @BenchmarkMode(Mode.SingleShotTime)
 @Fork(value=25, jvmArgsAppend={"-Xmx5g", "-Xms5g", "-Xmn3g", "-XX:+AlwaysPreTouch"})
 @OutputTimeUnit(TimeUnit.MILLISECONDS)
 @State(Scope.Benchmark)
-public class DifferentObjectSizesHashMap {
+public class AllLive {
 
     /*
-     * Test the System GC when 2/3 of the objects are live
-     * and kept reachable through a HashMap.
+     * Test the System GC when all allocated objects are live.
      *
      * The jvmArgs are provided to avoid GCs during object creation.
      */
 
-    static HashMap<Integer, byte[]> largeMap;
+    static ArrayList<Object[]> holder;
 
     @Setup(Level.Iteration)
     public void generateGarbage() {
-        largeMap = GarbageGenerator.generateAndFillHashMap(false);
-        int numberOfObjects = largeMap.size();
-        // Removing a third of the objects and keeping a good
-        // distribution of sizes.
-        for (int i = 0; i < numberOfObjects; i++) {
-            if (i%3 == 0) {
-                largeMap.remove(i);
-            }
-        }
+        holder = GarbageGenerator.generateObjectArrays();
     }
 
     @Benchmark
