+++ test/micro/org/openjdk/bench/vm/gc/systemgc/DifferentObjectSizesHashMap.java
@@ -20,7 +20,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package org.openjdk.bench.vm.gc.system_gc;
+package org.openjdk.bench.vm.gc.systemgc;
 
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.BenchmarkMode;
@@ -32,26 +32,35 @@ import org.openjdk.jmh.annotations.Scope;
 import org.openjdk.jmh.annotations.Setup;
 import org.openjdk.jmh.annotations.State;
 
+import java.util.HashMap;
 import java.util.concurrent.TimeUnit;
 
 @BenchmarkMode(Mode.SingleShotTime)
 @Fork(value=25, jvmArgsAppend={"-Xmx5g", "-Xms5g", "-Xmn3g", "-XX:+AlwaysPreTouch"})
 @OutputTimeUnit(TimeUnit.MILLISECONDS)
 @State(Scope.Benchmark)
-public class OneBigObject {
+public class DifferentObjectSizesHashMap {
 
     /*
-     * Test the System GC when there is a single large object.
+     * Test the System GC when 2/3 of the objects are live
+     * and kept reachable through a HashMap.
      *
-     * The heap settings provided are the same as for the other
-     * test for consistency.
+     * The jvmArgs are provided to avoid GCs during object creation.
      */
 
-    static Object[] holder;
+    static HashMap<Integer, byte[]> largeMap;
 
     @Setup(Level.Iteration)
     public void generateGarbage() {
-        holder = new Object[1024*1024*128];
+        largeMap = GarbageGenerator.generateAndFillHashMap(false);
+        int numberOfObjects = largeMap.size();
+        // Removing a third of the objects and keeping a good
+        // distribution of sizes.
+        for (int i = 0; i < numberOfObjects; i++) {
+            if (i%3 == 0) {
+                largeMap.remove(i);
+            }
+        }
     }
 
     @Benchmark
