+++ test/micro/org/openjdk/bench/vm/gc/systemgc/HalfHashedHalfDead.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, 2024, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -20,7 +20,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package org.openjdk.bench.vm.gc.system_gc;
+package org.openjdk.bench.vm.gc.systemgc;
 
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.BenchmarkMode;
@@ -39,11 +39,11 @@ import java.util.concurrent.TimeUnit;
 @Fork(value=25, jvmArgsAppend={"-Xmx5g", "-Xms5g", "-Xmn3g", "-XX:+AlwaysPreTouch"})
 @OutputTimeUnit(TimeUnit.MILLISECONDS)
 @State(Scope.Benchmark)
-public class HalfDeadFirstPart {
+public class HalfHashedHalfDead {
 
     /*
-     * Test the System GC when half of the objects are dead.
-     * In this test the first half of the objects are cleared.
+     * Test the System GC when there is a big amount of objects
+     * with hash codes calculated.
      *
      * The jvmArgs are provided to avoid GCs during object creation.
      */
@@ -53,9 +53,16 @@ public class HalfDeadFirstPart {
     @Setup(Level.Iteration)
     public void generateGarbage() {
         holder = GarbageGenerator.generateObjectArrays();
-        // Clearing every other object array in the holder
-        for (int i = 0; i < holder.size() / 2; i++) {
-            holder.set(i, null);
+        // Keeping half the objects and calculating the hash code to
+        // force some GCs to preserve marks.
+        for (Object[] objectArray: holder) {
+            for (int i = 0; i < objectArray.length; i++) {
+                if (i % 2 == 0) {
+                    objectArray[i].hashCode();
+                } else {
+                    objectArray[i] = null;
+                }
+            }
         }
     }
