+++ test/micro/org/openjdk/bench/vm/gc/systemgc/OneBigObject.java
@@ -20,7 +20,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package org.openjdk.bench.vm.gc.system_gc;
+package org.openjdk.bench.vm.gc.systemgc;
 
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.BenchmarkMode;
@@ -32,31 +32,26 @@ import org.openjdk.jmh.annotations.Scope;
 import org.openjdk.jmh.annotations.Setup;
 import org.openjdk.jmh.annotations.State;
 
-import java.util.ArrayList;
 import java.util.concurrent.TimeUnit;
 
 @BenchmarkMode(Mode.SingleShotTime)
 @Fork(value=25, jvmArgsAppend={"-Xmx5g", "-Xms5g", "-Xmn3g", "-XX:+AlwaysPreTouch"})
 @OutputTimeUnit(TimeUnit.MILLISECONDS)
 @State(Scope.Benchmark)
-public class HalfDeadSecondPart {
+public class OneBigObject {
 
     /*
-     * Test the System GC when half of the objects are dead.
-     * In this test the second half of the objects are cleared.
+     * Test the System GC when there is a single large object.
      *
-     * The jvmArgs are provided to avoid GCs during object creation.
+     * The heap settings provided are the same as for the other
+     * test for consistency.
      */
 
-    static ArrayList<Object[]> holder;
+    static Object[] holder;
 
     @Setup(Level.Iteration)
     public void generateGarbage() {
-        holder = GarbageGenerator.generateObjectArrays();
-        // Clearing every other object array in the holder
-        for (int i = holder.size() / 2; i < holder.size(); i++) {
-            holder.set(i, null);
-        }
+        holder = new Object[1024*1024*128];
     }
 
     @Benchmark
